Add styling to form
