+++ src/style.css
@@ -65,3 +65,18 @@ input[type="checkbox"] {
   color: #999;
   border-color: #999;
 }
+#edit{
+    position: relative;
+
+}
+#delete{
+    position: relative;
+}
+#identifier{
+    display: flex;
+    justify-content: space-between;
+}
+form{
+    display: flex;
+    justify-content: center;
+}
